refactor(utils): fix tests and move constants to utils file
