+++ index.js
@@ -1,3 +1,4 @@
+import { CSS_DOCS_URL, TECH_DOCS_URL } from './utils.js'
 import rules from './rules.js'
 import {
   variantBreakpoints,
@@ -19,8 +20,8 @@ const variants = [
 ];
 
 export const presetMigrate = () => {
-  console.log("Read more about supported classes in Warp CSS Docs: https://warp-ds.github.io/css-docs/")
-  console.log("You can find currently supported components in Warp Tech Docs: https://warp-ds.github.io/tech-docs")
+  console.log(`Read more about supported classes in Warp CSS Docs: ${CSS_DOCS_URL}`)
+  console.log(`You can find currently supported components in Warp Tech Docs: ${TECH_DOCS_URL}`)
 
   return { name: '@warp-ds/preset-migrate', theme, rules, variants }
 }
